+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 --------------------
 - Chg: Remove `'sort' => 'older'` for the "Spaces" top menu entry
 - Enh: Added `transition/embed/mail-conversation-sidebar`
+- Enh: Added `transition/embed/notification-overview`
 
 1.7.1 (Nov 27, 2023)
 --------------------